requirements.txt
Flask~=1.1.2
python-dotenv~=0.16.0
SQLAlchemy~=1.4.3